Remove excessive optimizing code.

+++ src/org/hlib4j/collection/Multiple.java
@@ -295,7 +295,7 @@ public class Multiple<E extends Number> implements Rule<E>
 	{
 		try
 		{
-			return (anAtomicInteger.intValue() % this.refValue.intValue()) == 0;
+			return (anAtomicInteger.intValue() % ((AtomicInteger) this.refValue).intValue()) == 0;
 		}
 		catch (ClassCastException e)
 		{
@@ -316,7 +316,7 @@ public class Multiple<E extends Number> implements Rule<E>
 	{
 		try
 		{
-			return (anAtomicLong.longValue() % this.refValue.longValue()) == 0L;
+			return (anAtomicLong.longValue() % ((AtomicLong) this.refValue).longValue()) == 0L;
 		}
 		catch (ClassCastException e)
 		{
